Non working particle shader, some interference with the regular shader.

+++ src/Particle.h
@@ -9,6 +9,6 @@ struct Particle
 	vec3 pos;
 	vec3 velocity;
 	float invmass;
-	int phase;
+	float phase;
 };
